+++ tests/Core/Serialization/DOM/Normalizers/MetadataNormalizerTest.php
@@ -32,6 +32,7 @@ use CycloneDX\Core\Serialization\DOM\Normalizers;
 use CycloneDX\Core\Spec\Spec;
 use CycloneDX\Tests\_traits\DomNodeAssertionTrait;
 use DateTime;
+use DateTimeZone;
 use DomainException;
 use DOMDocument;
 use PHPUnit\Framework\TestCase;
@@ -59,10 +60,10 @@ class MetadataNormalizerTest extends TestCase
         self::assertStringEqualsDomNode('<metadata></metadata>', $actual);
     }
 
-    public function testNormalizeTimestamp(): void
+    public function testNormalizeTimestampIsToZuluWithoutModifications(): void
     {
-        $fakeDate = 'just-now';
-        $timestamp = $this->createMock(DateTime::class);
+        $timeZone = new DateTimeZone('Pacific/Nauru');
+        $timestamp = new DateTime('2000-01-01 00:00:00', $timeZone);
         $metadata = $this->createConfiguredMock(
             Metadata::class,
             ['getTimestamp' => $timestamp]
@@ -77,15 +78,17 @@ class MetadataNormalizerTest extends TestCase
         );
         $normalizer = new Normalizers\MetadataNormalizer($factory);
 
-        $timestamp->method('format')
-            ->with('c')
-            ->willReturn($fakeDate);
-
         $actual = $normalizer->normalize($metadata);
 
         self::assertStringEqualsDomNode(
-            "<metadata><timestamp>$fakeDate</timestamp></metadata>",
-            $actual
+            '<metadata><timestamp>1999-12-31T12:00:00Z</timestamp></metadata>',
+            $actual,
+            'not the expected Zulu time'
+        );
+        self::assertSame(
+            '2000-01-01T00:00:00+12:00',
+            $timestamp->format('c'),
+            'timestamp was modified'
         );
     }